refactor: Change export memory
