+++ src/services/auth/index.ts
@@ -1,3 +1,2 @@
 export * from './auth.declaration';
-export * from './auth.memory';
 export * from './auth';
