+++ src/services/http/index.ts
@@ -1,3 +1,2 @@
 export * from './http.declaration';
-export * from './http.memory';
 export * from './http';
